# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders": {
      "get": {
        "operationId": "OrderController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "订单列表"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询当前用户订单列表",
        "tags": [
          "order"
        ]
      },
      "post": {
        "operationId": "OrderController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "订单已创建"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建订单（基于账单，金额校验）",
        "tags": [
          "order"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrderController_findById",
        "parameters": [],
        "responses": {
          "200": {
            "description": "订单信息"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "按 id 查询订单",
        "tags": [
          "order"
        ]
      }
    },
    "/orders/{id}/submit-payment": {
      "patch": {
        "operationId": "OrderController_submitPayment",
        "parameters": [],
        "responses": {
          "200": {
            "description": "已进入待支付"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "提交支付（CREATED → PENDING_PAYMENT）",
        "tags": [
          "order"
        ]
      }
    },
    "/orders/{id}/confirm-paid": {
      "patch": {
        "operationId": "OrderController_confirmPaid",
        "parameters": [],
        "responses": {
          "200": {
            "description": "支付成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "确认支付成功（PENDING_PAYMENT → PAID，联动账单）",
        "tags": [
          "order"
        ]
      }
    },
    "/orders/{id}/complete": {
      "patch": {
        "operationId": "OrderController_complete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "订单已完成"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "完成订单（PAID → COMPLETED）",
        "tags": [
          "order"
        ]
      }
    },
    "/orders/{id}/close": {
      "patch": {
        "operationId": "OrderController_close",
        "parameters": [],
        "responses": {
          "200": {
            "description": "订单已关闭"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "关闭订单（超时/取消 → CLOSED）",
        "tags": [
          "order"
        ]
      }
    },
    "/payments": {
      "get": {
        "operationId": "PaymentController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "支付流水列表"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询当前用户支付流水",
        "tags": [
          "payment"
        ]
      },
      "post": {
        "operationId": "PaymentController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "支付已创建"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建支付（下单）",
        "tags": [
          "payment"
        ]
      }
    },
    "/payments/{id}/mock-complete": {
      "post": {
        "operationId": "PaymentController_completeMockPayment",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Mock 回调已处理"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "开发环境完成 Mock 支付并触发已签名回调",
        "tags": [
          "payment"
        ]
      }
    },
...
    "/payments/{id}": {
      "get": {
        "operationId": "PaymentController_findById",
        "parameters": [],
        "responses": {
          "200": {
            "description": "支付流水信息"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询支付流水",
        "tags": [
          "payment"
        ]
      }
    },
    "/payments/{id}/close": {
      "patch": {
        "operationId": "PaymentController_close",
        "parameters": [],
        "responses": {
          "200": {
            "description": "支付已关闭"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "关闭支付",
        "tags": [
          "payment"
        ]
      }
    },
    "/payments/{id}/refund": {
      "post": {
        "operationId": "PaymentController_refund",
        "parameters": [],
        "responses": {
          "201": {
            "description": "退款单已创建"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "发起退款",
        "tags": [
          "payment"
        ]
      }
    },
    "/payments/reconcile/daily": {
      "post": {
        "operationId": "PaymentController_reconcile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "对账结果"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "执行日对账",
        "tags": [
          "payment"
        ]
      }
    },
...
    "/payments/reconcile/diff/{id}/resolve": {
      "patch": {
        "operationId": "PaymentController_resolveDiff",
        "parameters": [],
        "responses": {
          "200": {
            "description": "差异已解决"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "解决对账差异（PENDING → RESOLVED，解锁冻结退款）",
        "tags": [
          "payment"
        ]
      }
    },
...
  },
  "info": {
    "title": "Helio API",
    "description": "Helio 太阳能能源监控平台 API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}